fix(data-table): move EditableCell to web app data-table-content
- Remove EditableCell from packages/ui (wrong table), revert body row to plain flexRender
- Add EditableCell + wiring to apps/web data-table-content (correct table)
- Add pencil icon + ring hover affordance to make editable cells clearly clickable
